Conflicts:
	Controller/MainController.cpp
	Windows/CreateProcess.cpp
	wxWidgets/Controller/wxDynLibCPUcontroller.cpp

+++ Windows/PowerProf/PowerProfDynLinked.cpp
@@ -117,6 +117,9 @@ bool PowerProfDynLinked::DisableFrequencyScalingByOS()
   }
   if( bDesiredPowerSchemeExists )
   {
+    //TODO (Windows XP): it seems that the DVFS can not always be disabled:
+    // when the service runs at "LocalSystem" then an administrator's power
+    // scheme may override the LocalSystem's scheme.
     mp_i_powerprofdynlinked->DisableDVFSforPowerSchemeToSet() ;
     mp_i_powerprofdynlinked->ActivatePowerSchemeToSet() ;
   }
@@ -177,7 +180,8 @@ void PowerProfDynLinked::OutputAllPowerSchemes()
   }
 }
 
-BYTE PowerProfDynLinked::SetActivePowerScheme(
+//BYTE
+DWORD PowerProfDynLinked::SetActivePowerScheme(
   const std::wstring & r_stdwstrPowerSchemeName )
 {
   return mp_i_powerprofdynlinked->SetActivePowerScheme(
